#149: move crs modules from core.io.ini.models to core.io.crosssection.models namespace.
