+++ docs/reference/crosssection.md
@@ -1,0 +1,8 @@
+## cross section .ini files
+
+The crosssection module provides the specific logic for accessing cross section files (location and definition). Generic parsing and serializing functionality comes from the generic hydrolib.core.io.ini modules.
+
+
+### Model
+
+::: hydrolib.core.io.crosssection.models
