+++ hydrolib/core/io/mdu/models.py
@@ -3,15 +3,10 @@ from typing import Dict, List, Literal, Optional, Union
 
 from pydantic import Field
 
+from hydrolib.core.io.crosssection.models import CrossDefModel, CrossLocModel
 from hydrolib.core.io.ext.models import ExtModel
 from hydrolib.core.io.friction.models import FrictionModel
-from hydrolib.core.io.ini.models import (
-    CrossDefModel,
-    CrossLocModel,
-    INIBasedModel,
-    INIGeneral,
-    INIModel,
-)
+from hydrolib.core.io.ini.models import INIBasedModel, INIGeneral, INIModel
 from hydrolib.core.io.ini.util import get_split_string_on_delimiter_validator
 from hydrolib.core.io.net.models import NetworkModel
 from hydrolib.core.io.polyfile.models import PolyFile
